refactor(web): group client runtime modules
